Updated to support AUTOINCREMENT INTEGER ID
If user supplied s static $_autoId = true;  variable in their class, then it foregoes the char(20) code generated GUID id, and also grabs the last inserted ID after a successful insert and populates the object with that ID. I modelled this behaviour after the static $_table property.  So a user can keep using this library as intended and never take advantage of this or they can, if they want.